issue #139 refactor: move getCurrentWeek directory to calendarUtils and change import statement
